updated index.html with corrections
The bit about the *shx file wasn't quite correct, and I thought a little misleading. There are other indexing sidecar files that Esri uses, but those aren't part of the universal spec, so I didn't mention those.

I edited the section "Multifile format" because it contained an error about which 3 files are required.